requirements.txt
openai>=1.16.0
pydantic>=2.7.0
ffmpeg-python==0.2.0
typing-extensions>=4.9.0
python-dotenv==1.0.0

# Type checking
mypy==1.8.0

# Linting
pylint==3.0.3
flake8==7.0.0
black==24.1.1
isort==5.13.2 
